refactor(projects): split projects-view monolith into a component tree
Behavior-preserving split of the 1,400-line projects-view.tsx into a small,
navigable tree — the Phase-0 goal from the redesign plan, deferred until now
because the source-text test pinned everything to the one file.

  src/components/projects-view.tsx        container (ProjectsView + the move
                                          undo toast + data/DnD/keyboard wiring)
  src/components/projects/project-row.tsx ProjectRow (a project card)
  src/components/projects/session-row.tsx ProjectChatRow (a session row)
  src/components/projects/projects-shared.ts  shared pure helpers + MoveTarget
                                          (chatDotClass, lastActiveMs, shortRoot,
                                          openSessionById, CHAT_CAP)

No behavior change: the public import (@/components/projects-view → ProjectsView)
is unchanged, and no JSX/logic was modified — only relocated. Import graph is
acyclic (container → project-row → session-row → shared).

Test-architecture change that unblocks this: projects-view.test.ts now reads the
four files and runs its ~85 source-text assertions against the COMBINED source,
so each keeps verifying behavior regardless of which file a piece lives in.
